+++ frontend/src/components/@commons/image/Image.stories.tsx
@@ -1,5 +1,5 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
-import Image from 'components/@commons/Image';
+import Image from 'components/@commons/image/Image';
 
 export default {
   title: 'Image',
