+++ frontend/src/components/interviewQuestion/InterviewQuestionSearchForm.tsx
@@ -5,8 +5,8 @@ import useInterviewQuestionSearch from 'hooks/questionSearch/useInterviewQuestio
 import deleteIcon from 'assets/images/close.svg';
 import searchIcon from 'assets/images/search.svg';
 
-import Image from 'components/@commons/Image';
 import Button from 'components/@commons/button/Button';
+import Image from 'components/@commons/image/Image';
 import Input from 'components/@commons/input/Input';
 
 const InterviewQuestionSearchForm = () => {
